Refactor telemetry mocks for improved testing
Remove obsolete MockTelemetryItem and MockSpan classes, replacing them with MockTelemetryItem and MockSpanV2 for enhanced testing capabilities. Update relevant tests to utilize the new mock classes, ensuring better coverage and reliability in telemetry processing tests.
